requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via starlette
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
exceptiongroup==1.1.3
    # via anyio
fastapi==0.102.0
    # via -r requirements.in
ffmpeg-python==0.2.0
    # via -r requirements.in
future==0.18.3
    # via ffmpeg-python
idna==3.4
    # via
    #   anyio
    #   requests
pydantic==2.3.0
    # via fastapi
pydantic-core==2.6.3
    # via pydantic
requests==2.31.0
    # via
    #   speechrecognition
    #   vadersentiment
sniffio==1.3.0
    # via anyio
speechrecognition==3.10.0
    # via -r requirements.in
starlette==0.27.0
    # via fastapi
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.0.4
    # via requests
vadersentiment==3.3.2
    # via -r requirements.in
